+++ src/Contracts/PaginationStrategyInterface.php
@@ -47,4 +47,3 @@ interface PaginationStrategyInterface
      */
     public function excludedRequestParameters(): array;
 }
-